Validations: DELETE exceptions!

+++ src/main/java/com/devsuperior/desafioCrudClientes/services/ClientService.java
@@ -47,6 +47,9 @@ public class ClientService {
 
     @Transactional
     public void delete(Long id){
+        if(!clientRepository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
         clientRepository.deleteById(id);
     }
